Обработка некорректных rest-запросов

+++ src/main/java/com/company/controller/rest/AccountController.java
@@ -4,6 +4,8 @@ import com.company.dao.AccountDAO;
 import com.company.domain.AccountDomain;
 import com.company.domain.SecureAccountDomain;
 import com.company.entity.Account;
+import org.hibernate.StaleStateException;
+import org.hibernate.exception.ConstraintViolationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.HttpStatus;
@@ -11,19 +13,28 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
-import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
 import java.util.List;
 
+/**
+ * Контроллер, управляющий частью Rest-сервиса.
+ * Здесь представлены методы управления аккаунтами пользователей
+ */
 @RestController
 @RequestMapping(value = "/account")
 public class AccountController {
+    private static final String BAD_REQ = "Bad Request: ";
 
     @Autowired
     @Qualifier("accountDAO")
     AccountDAO accountDAO;
 
-    //@todo расписать варианты кодов ошибок
+    /**
+     * Возвращает информацию об аккаунте аутентифицированного пользователя
+     *
+     * @param authentication
+     * @return AccountDomain и HttpStatus.OK, если успех
+     */
     @RequestMapping(value = "/", method = RequestMethod.GET)
     public ResponseEntity<AccountDomain> getCurrentAccount(Authentication authentication) {
         String auth = authentication.getName();
@@ -31,12 +42,27 @@ public class AccountController {
         return new ResponseEntity<>(accountDomain, HttpStatus.OK);
     }
 
+    /**
+     * Возвращает информацию об аккаунте пользователя с id
+     *
+     * @param accountId id аккаунта в базе данных
+     * @return AccountDomain и HttpStatus.OK, если успех,
+     * если не найден пользователь с id accountId, HttpStatus.BAD_REQUEST
+     */
     @RequestMapping(value = "/{accountId}", method = RequestMethod.GET)
     public ResponseEntity<AccountDomain> getAccount(@PathVariable int accountId) {
         Account account = accountDAO.read(accountId);
+        if (account == null) {
+            return new ResponseEntity(BAD_REQ + "Account with id " + accountId + " doesn't exists", HttpStatus.BAD_REQUEST);
+        }
         return new ResponseEntity<>(new AccountDomain(account), HttpStatus.OK);
     }
 
+    /**
+     * Возвращает информацию обо всех пользовательских аккаунтах
+     *
+     * @return List<AccountDomain> и HttpStatus.OK, если успех
+     */
     @RequestMapping(value = "/all", method = RequestMethod.GET)
     public ResponseEntity<List<AccountDomain>> getAllAccounts() {
         List<AccountDomain> accountDomains = new ArrayList<>();
@@ -46,16 +72,46 @@ public class AccountController {
         return new ResponseEntity<>(accountDomains, HttpStatus.OK);
     }
 
+    /**
+     * Создает аккаунт пользователя в системе
+     *
+     * @param accountDomain бин-эквивалент Account сущности
+     * @return HttpStatus.OK, если успех
+     * если создается аккаунт с неуникальным именем, HttpStatus.BAD_REQUEST
+     */
     @RequestMapping(value = "/", method = RequestMethod.POST)
-    public ResponseEntity createAccount(@RequestBody SecureAccountDomain accountDomain) throws NoSuchAlgorithmException {
+    public ResponseEntity createAccount(@RequestBody SecureAccountDomain accountDomain) {
         accountDomain.encodePass();
-        accountDAO.create(accountDomain);
+        try {
+            accountDAO.create(accountDomain);
+        } catch (ConstraintViolationException exc) {
+            return new ResponseEntity(BAD_REQ + exc.getClass().getName()
+                    + ". Probably, your nickname " + accountDomain.getNickname() + " is not unique", HttpStatus.BAD_REQUEST);
+        }
         return new ResponseEntity(HttpStatus.OK);
     }
 
+    /**
+     * Удаляет аккаунт пользователя из системы
+     *
+     * @param accountId id аккаунта в базе данных
+     * @return HttpStatus.OK, если успех
+     * если удаляется текущий аутентифицированный пользователь, HttpStatus.BAD_REQUEST
+     * если удаляется пользователь с неизвестным id, HttpStatus.BAD_REQUEST
+     */
     @RequestMapping(value = "/{accountId}", method = RequestMethod.DELETE)
-    public ResponseEntity deleteAccount(@PathVariable int accountId) {
-        accountDAO.delete(accountId);
+    public ResponseEntity deleteAccount(@PathVariable int accountId, Authentication authentication) {
+        String auth = authentication.getName();
+        Account account = accountDAO.read(auth);
+        if (account.getId() == accountId) {
+            return new ResponseEntity(BAD_REQ + "SelfDeletion is not allowed", HttpStatus.BAD_REQUEST);
+        }
+        try {
+            accountDAO.delete(accountId);
+        } catch (StaleStateException exc) {
+            return new ResponseEntity(BAD_REQ + exc.getClass()
+                    + ". Probably, you are trying to delete non-existent account", HttpStatus.BAD_REQUEST);
+        }
         return new ResponseEntity(HttpStatus.OK);
     }
 }
